Correct misuse of BumpyMatrix::splitAsBumpyMatrix()
BumpyMatrix::splitAsBumpyMatrix() was misued in the SpatialExperiment
examples, which resulted in the returned 'mol' object to be an
incorrect BumpyMatrix object. This caused the call to SpatialExperiment()
to fail further down (using SummarizedExperiment 1.23.2, the latest version)
because the dimnames of second supplied assay 'mol' didn't align with those
of first supplied assay 'y' (this is now a requirement).